+++ src/options.js
@@ -4,14 +4,17 @@ const { OptionValueInvalidError } = require('./errors.js')
 // All setters expect to get a string
 class Options {
   constructor () {
-    this.options = Object.assign({}, this.defaults)
+    this._options = Object.assign({}, this.defaults)
 
-    const keys = Object.keys(this.options)
+    // We'll use the keys from the defaults
+    // to proxy get calls directly to the
+    // _options object
+    const keys = Object.keys(this.defaults)
 
     return new Proxy(this, {
       get (target, name) {
         if (keys.indexOf(name) !== -1) {
-          return Reflect.get(target, 'options')[name]
+          return Reflect.get(target, '_options')[name]
         } else {
           return target[name]
         }
@@ -23,13 +26,21 @@ class Options {
     })
   }
 
-  updateBackgroundRGB () {
-    let color = Color(this.options.backgroundcolor).alpha(this.options.backgroundalpha)
-    this.options.backgroundrgb = color.rgb().string()
+  hasOption (key) {
+    return key in this._options
+  }
+
+  _updateBackgroundRGB () {
+    const color = Color(this._options.backgroundcolor).alpha(this._options.backgroundalpha)
+    this._options.backgroundrgb = color.rgb().string()
+  }
+
+  get options () {
+    return Object.freeze(Object.assign({}, this._options))
   }
 
   get defaults () {
-    return {
+    return Object.freeze({
       backgroundalpha: 1, // 0-1
       backgroundcolor: '#000', // hex without the #
       backgroundrgb: 'rgb(0, 0, 0)', // automatically updated when color or alpha is set
@@ -45,77 +56,77 @@ class Options {
       top: null, // number (used for custom extraction/crop)
       version: null, // anything
       width: null // auto/aspect ratio by default
-    }
+    })
   }
 
   set backgroundalpha (value) {
-    this.options.backgroundalpha = parseFloat(value)
-    this.updateBackgroundRGB()
+    this._options.backgroundalpha = parseFloat(value)
+    this._updateBackgroundRGB()
   }
 
   set backgroundcolor (value) {
-    this.options.backgroundcolor = `#${value}`
-    this.updateBackgroundRGB()
+    this._options.backgroundcolor = `#${value}`
+    this._updateBackgroundRGB()
   }
 
   set blur (value) {
     value = parseFloat(value)
     if (value < 0.3 || value > 1000) throw new OptionValueInvalidError(`'${value}' is an invalid value for option 'blur'`)
-    this.options.blur = value
+    this._options.blur = value
   }
 
   set fit (value) {
-    let valid = ['cover', 'contain', 'fill', 'inside', 'outside']
+    const valid = ['cover', 'contain', 'fill', 'inside', 'outside']
     if (valid.indexOf(value) === -1) throw new OptionValueInvalidError(`'${value}' is an invalid value for option 'fit'`)
-    this.options.fit = value
+    this._options.fit = value
   }
 
   set gravity (value) {
-    let valid = ['north', 'northeast', 'east', 'southeast', 'south', 'southwest', 'west', 'northwest', 'center', 'entropy', 'attention']
+    const valid = ['north', 'northeast', 'east', 'southeast', 'south', 'southwest', 'west', 'northwest', 'center', 'entropy', 'attention']
     if (valid.indexOf(value) === -1) throw new OptionValueInvalidError(`'${value}' is an invalid value for option 'gravity'`)
-    this.options.gravity = value
+    this._options.gravity = value
   }
 
   set height (value) {
-    this.options.height = parseFloat(value)
+    this._options.height = parseFloat(value)
   }
 
   set left (value) {
-    this.options.left = parseFloat(value)
+    this._options.left = parseFloat(value)
   }
 
   set mirror (value) {
-    let valid = ['x', 'y']
+    const valid = ['x', 'y']
     if (valid.indexOf(value) === -1) throw new OptionValueInvalidError(`'${value}' is an invalid value for option 'mirror'`)
-    this.options.mirror = value
+    this._options.mirror = value
   }
 
   set output (value) {
-    let valid = ['jpeg', 'png', 'webp']
+    const valid = ['jpeg', 'png', 'webp']
     if (valid.indexOf(value) === -1) throw new OptionValueInvalidError(`'${value}' is an invalid value for option 'output'`)
-    this.options.output = value
+    this._options.output = value
   }
 
   set quality (value) {
     value = parseInt(value, 10)
     if (value < 1 || value > 100) throw new OptionValueInvalidError(`'${value}' is an invalid value for option 'quality'`)
-    this.options.quality = value
+    this._options.quality = value
   }
 
   set rotate (value) {
-    this.options.rotate = parseFloat(value)
+    this._options.rotate = parseFloat(value)
   }
 
   set top (value) {
-    this.options.top = parseFloat(value)
+    this._options.top = parseFloat(value)
   }
 
   set version (value) {
-    this.options.version = value
+    this._options.version = value
   }
 
   set width (value) {
-    this.options.width = parseFloat(value)
+    this._options.width = parseFloat(value)
   }
 }