+++ src/parser.test.js
@@ -4,32 +4,32 @@ const Parser = require('./parser.js')
 const crypto = require('crypto')
 
 function sign (token, options, path) {
-  let hmac = crypto.createHmac('sha1', token)
+  const hmac = crypto.createHmac('sha1', token)
   hmac.update(`${options}/${path}`)
-  let signature = hmac.digest('hex')
+  const signature = hmac.digest('hex')
   return `${options};signature:${signature}/${path}`
 }
 
-let path = 'path/to/image.jpg'
+const path = 'path/to/image.jpg'
 
 test('path works with leading /', () => {
-  let options = 'width:100'
-  let parser = new Parser(`/${options}/${path}`)
+  const options = 'width:100'
+  const parser = new Parser(`/${options}/${path}`)
   parser.parse()
   expect(parser.path).toBe(path)
   expect(parser.optionsString).toBe(options)
 })
 
 test('path works without leading /', () => {
-  let options = 'width:100'
-  let parser = new Parser(`${options}/${path}`)
+  const options = 'width:100'
+  const parser = new Parser(`${options}/${path}`)
   parser.parse()
   expect(parser.path).toBe(path)
   expect(parser.optionsString).toBe(options)
 })
 
 test('signature missing', () => {
-  let parser = new Parser(`width:100/${path}`, '123', true)
+  const parser = new Parser(`width:100/${path}`, '123', true)
   expect(parser.token).toBe('123')
   expect(parser.signatureRequired).toBe(true)
   expect(() => { parser.parse() }).toThrow(SignatureMissingError)
@@ -37,87 +37,93 @@ test('signature missing', () => {
 })
 
 test('signature mismatch', () => {
-  let uri = sign('456', 'width:100', path)
-  let parser = new Parser(uri, '123', true)
+  const uri = sign('456', 'width:100', path)
+  const parser = new Parser(uri, '123', true)
   expect(() => { parser.parse() }).not.toThrow(SignatureMissingError)
   expect(() => { parser.parse() }).toThrow(SignatureMismatchError)
 })
 
 test('signature match', () => {
-  let uri = sign('123', 'width:100', path)
-  let parser = new Parser(uri, '123', true)
+  const uri = sign('123', 'width:100', path)
+  const parser = new Parser(uri, '123', true)
   expect(() => { parser.parse() }).not.toThrow(SignatureMissingError)
   expect(() => { parser.parse() }).not.toThrow(SignatureMismatchError)
 })
 
 test('unknown option key', () => {
-  let parser = new Parser(`notakey:notavalue/${path}`)
+  const parser = new Parser(`notakey:notavalue/${path}`)
   expect(() => { parser.parse() }).toThrow(OptionKeyUnknownError)
 })
 
 test('invalid option value', () => {
-  let parser = new Parser(`blur:0/${path}`)
+  const parser = new Parser(`blur:0/${path}`)
   expect(() => { parser.parse() }).toThrow(OptionValueInvalidError)
 })
 
+test('optionsCollection is a collection of frozen objects', () => {
+  const parser = new Parser(`width:100;blur:0.3/${path}`)
+  const parsedOptionsCollection = parser.parse()
+  expect(parsedOptionsCollection.length).toBe(2)
+  expect(Object.isFrozen(parsedOptionsCollection[0])).toBe(true)
+  expect(Object.isFrozen(parsedOptionsCollection[1])).toBe(true)
+})
+
 test('multiple options without signature', () => {
-  let options = [
+  const options = [
     ['width', '100'],
     ['height', '100'],
     ['output', 'webp']
   ]
-  let expected = [
+  const expected = [
     ['width', 100],
     ['height', 100],
     ['output', 'webp']
   ]
 
-  let delimiters = ['_', ':', '=']
+  const delimiters = ['_', ':', '=']
 
   delimiters.forEach((delimiter) => {
-    let optionsString = options.map((option) => option.join(delimiter)).join(',')
+    const optionsString = options.map((option) => option.join(delimiter)).join(',')
 
-    let uri = `${optionsString}/${path}`
-    let parser = new Parser(uri)
-    let parsedOptionsCollection = parser.parse()[0]
+    const uri = `${optionsString}/${path}`
+    const parser = new Parser(uri)
+    const parsedOptionsCollection = parser.parse()[0]
 
     expected.forEach((expectedSetting) => {
-      expect(parsedOptionsCollection.options[expectedSetting[0]]).toBe(expectedSetting[1])
       expect(parsedOptionsCollection[expectedSetting[0]]).toBe(expectedSetting[1])
     })
   })
 })
 
 test('multiple options with signature', () => {
-  let options = [
+  const options = [
     ['width', '100'],
     ['height', '100'],
     ['output', 'webp']
   ]
-  let expected = [
+  const expected = [
     ['width', 100],
     ['height', 100],
     ['output', 'webp']
   ]
 
-  let delimiters = ['_', ':', '=']
+  const delimiters = ['_', ':', '=']
 
   delimiters.forEach((delimiter) => {
-    let optionsString = options.map((option) => option.join(delimiter)).join(',')
+    const optionsString = options.map((option) => option.join(delimiter)).join(',')
 
-    let uri = sign('123', optionsString, path)
-    let parser = new Parser(uri, '123', true)
-    let parsedOptionsCollection = parser.parse()[0]
+    const uri = sign('123', optionsString, path)
+    const parser = new Parser(uri, '123', true)
+    const parsedOptionsCollection = parser.parse()[0]
 
     expected.forEach((expectedSetting) => {
-      expect(parsedOptionsCollection.options[expectedSetting[0]]).toBe(expectedSetting[1])
       expect(parsedOptionsCollection[expectedSetting[0]]).toBe(expectedSetting[1])
     })
   })
 })
 
 test('compound options without signature', () => {
-  let optionsCollection = [
+  const optionsCollection = [
     [
       ['width', '100'],
       ['height', '100']
@@ -127,7 +133,7 @@ test('compound options without signature', () => {
       ['output', 'webp']
     ]
   ]
-  let expectedCollection = [
+  const expectedCollection = [
     [
       ['width', 100],
       ['height', 100]
@@ -138,18 +144,17 @@ test('compound options without signature', () => {
     ]
   ]
 
-  let delimiters = ['_', ':', '=']
+  const delimiters = ['_', ':', '=']
 
   delimiters.forEach((delimiter) => {
-    let options = optionsCollection.map((options) => {
+    const options = optionsCollection.map((options) => {
       return options.map((option) => option.join('_')).join(',')
     }).join(';')
 
-    let parser = new Parser(`${options}/${path}`)
-    let parsedOptionsCollection = parser.parse()
+    const parser = new Parser(`${options}/${path}`)
+    const parsedOptionsCollection = parser.parse()
     parsedOptionsCollection.forEach((parsedOptions, index) => {
       expectedCollection[index].forEach((expectedSetting) => {
-        expect(parsedOptions.options[expectedSetting[0]]).toBe(expectedSetting[1])
         expect(parsedOptions[expectedSetting[0]]).toBe(expectedSetting[1])
       })
     })
@@ -157,7 +162,7 @@ test('compound options without signature', () => {
 })
 
 test('compound options with signature', () => {
-  let optionsCollection = [
+  const optionsCollection = [
     [
       ['width', '100'],
       ['height', '100']
@@ -167,7 +172,7 @@ test('compound options with signature', () => {
       ['output', 'webp']
     ]
   ]
-  let expectedCollection = [
+  const expectedCollection = [
     [
       ['width', 100],
       ['height', 100]
@@ -178,19 +183,18 @@ test('compound options with signature', () => {
     ]
   ]
 
-  let delimiters = ['_', ':', '=']
+  const delimiters = ['_', ':', '=']
 
   delimiters.forEach((delimiter) => {
-    let options = optionsCollection.map((options) => {
+    const options = optionsCollection.map((options) => {
       return options.map((option) => option.join('_')).join(',')
     }).join(';')
 
-    let uri = sign('123', options, path)
-    let parser = new Parser(uri, '123', true)
-    let parsedOptionsCollection = parser.parse()
+    const uri = sign('123', options, path)
+    const parser = new Parser(uri, '123', true)
+    const parsedOptionsCollection = parser.parse()
     parsedOptionsCollection.forEach((parsedOptions, index) => {
       expectedCollection[index].forEach((expectedSetting) => {
-        expect(parsedOptions.options[expectedSetting[0]]).toBe(expectedSetting[1])
         expect(parsedOptions[expectedSetting[0]]).toBe(expectedSetting[1])
       })
     })